Split the held item renderer up into a magical and non-magical variant since they not neccessarily neet to be together
